type change

+++ agents/ravine.py
@@ -1,3 +1,3 @@
 class Ravine():
-    def __init__(self, skill_required: int):
+    def __init__(self, skill_required: float):
         self.skill_required = skill_required
